requirements.txt
langchain 
langchain 
langchain-groq 
langsmith 
python-dotenv
streamlit